Separate Branch Prediction IF from Simple Predictor, per review comments
