+++ core/SimpleBranchPred.hpp
@@ -1,27 +1,14 @@
-// <Branch.hpp> -*- C++ -*-
+// <SimpleBranchPred.hpp> -*- C++ -*-
 
 //!
-//! \file BranchPred.hpp
-//! \brief Definition of Branch Prediction API
+//! \file SimpleBranchPred.hpp
+//! \brief Class definition of a simple brranch prediction using branch prediction interface
 //!
 
 /*
- * This file defines the Branch Prediction API.
- * The goal is to define an API that is generic and yet flexible enough to support various
- * branch prediction microarchitecture.
- * To the end, we envision a generic branch predictor as a black box with following inputs
- * and outputs:
- *   * A generic Prediction output
- *   * A generic Prediction input
- *   * A generic Update input
- *
- * The generic branch predictor may have two operations:
- *   * getPrediction: produces Prediction output based on the Prediction input.
- *   * updatePredictor: updates Predictor with Update input.
- *
- * It is intended that an implementation of branch predictor must also specify
- * implementations of Prediction output, Prediction input and Update input, along with
- * implementations of getPrediction and updatePredictor operations.
+ * This file defines the class SimpleBranchPredictor, as well as, a default Prediction 
+ * output class, a default Prediction input class, a defeault Prediction input class
+ * as required by Olympia's Branch Prediction inteface
  * */
 #pragma once
 
@@ -29,20 +16,10 @@
 #include <map>
 #include <limits>
 #include "sparta/utils/SpartaAssert.hpp"
+#include "BranchPredIF.hpp"
 
 namespace olympia
 {
-
-    template <class PredictionT, class UpdateT, class InputT>
-    class BranchPredictorIF
-    {
-    public:
-        // TODO: create constexpr for bytes per compressed and uncompressed inst
-        static constexpr uint8_t bytes_per_inst = 4;
-        virtual ~BranchPredictorIF() { };
-        virtual PredictionT getPrediction(const InputT &) = 0;
-        virtual void updatePredictor(const UpdateT &) = 0;
-    };
 
     // following class definitions are example inputs & outputs for a very simple branch
     // predictor
