+++ test/core/branch_pred/BranchPred_test.cpp
@@ -1,4 +1,4 @@
-#include "BranchPred.hpp"
+#include "SimpleBranchPred.hpp"
 #include "sparta/utils/SpartaTester.hpp"
 
 TEST_INIT
